+++ postal_address/territory.py
@@ -9,17 +9,45 @@
 
 u""" Utilities to normalize and reconcile territory codes.
 
+.. data:: FOREIGN_TERRITORIES_MAPPING
+
+    Reference *valid* country_codes as keys that are foreign territories of
+    another country. The latter is the dictionary value.
+
 .. data:: COUNTRY_ALIASES
 
-.. data:: SUBDIVISION_ALIASES
+    Bind *invalid* country_code in the ISO-3166 meaning with their *valid*
+    iso counterpart.
+    This enables us to handle so special cases of commonly used country codes
+    that are not part of the ISO-3166 definitions.
+
+.. data:: SUBDIVISION_COUNTRIES
 
    Map subdivision ISO 3166-2 codes to their officially assigned ISO 3166-1
    alpha-2 country codes. Source: https://en.wikipedia.org/wiki
    /ISO_3166-2#Subdivisions_included_in_ISO_3166-1
 
+   In order to be able to build a parenting tree, this mapping should redirect
+   to the subdivision's closer country_code.
+
+.. data:: SUBDIVISION_ALIASES
+
+    Map some subdivision aliases representing the same territory, but defined
+    under different countries.
+
+.. data:: RESERVED_COUNTRY_CODES
+
+    Bind *invalid* country_code in the ISO-3166 (mostly subdivisions)
+    representing territories to *valid* iso country codes representing
+    the main country of this territory.
+
+.. data:: COUNTRY_ALIAS_TO_SUBDIVISION
+
+    Bind *invalid* country_code to their real subdivision code.
+
 .. data:: REVERSE_MAPPING
 
-   Reverse index of the SUBDIVISION_ALIASES mapping defined above.
+   Reverse index of the SUBDIVISION_COUNTRIES mapping defined above.
 """
 
 from __future__ import (
@@ -43,25 +71,52 @@ else:
     imap = map
     ifilter = filter
 
-
-COUNTRY_ALIASES = {
-    # Source:
-    # https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2#Exceptional_reservations
-    'AC': 'SH-AC',  # Ascension Island
-    'CP': 'FR-CP',  # Clipperton Island
-    'DG': 'IO',     # Diego Garcia
-    # 'EA': ['ES-ML', 'ES-CE'],  # Ceuta and Melilla
-    'FX': 'FR',     # France, Metropolitan
-    'IC': 'ES-CN',  # Canary Islands
-    'TA': 'SH-TA',  # Tristan da Cunha
-    # European Commision country code exceptions.
-    # Source: http://publications.europa.eu/code/pdf/370000en.htm#pays
-    'UK': 'GB',  # United Kingdom
-    'EL': 'GR',  # Greece
+FOREIGN_TERRITORIES_MAPPING = {
+    'CC': 'AU',  # Cocos Island,                      Australian territory
+    'HM': 'AU',  # Heard Island and McDonald Islands, Australian territory
+    'JE': 'BR',  # Jersey,                            Brazilian territory
+    'HK': 'CN',  # Hong Kong,                         Chinese territory
+    'MO': 'CN',  # Macao,                             Chinese territory
+    'FO': 'DK',  # Faroe Islands,                     Danish territory
+    'AX': 'FI',  # Åland,                             Finnish territory
+    'AQ': 'FR',  # Antarctica,                        French territory
+    'BL': 'FR',  # Saint Barthelemy,                  French territory
+    'GF': 'FR',  # French Guiana,                     French territory
+    'GP': 'FR',  # Guadeloupe,                        French territory
+    'GY': 'FR',  # Guyana,                            French territory
+    'MF': 'FR',  # Saint Martin,                      French territory
+    'MQ': 'FR',  # Martinique,                        French territory
+    'NC': 'FR',  # New Caledonia,                     French territory
+    'PF': 'FR',  # French Polynesia,                  French territory
+    'PM': 'FR',  # Saint Pierre and Miquelon,         French territory
+    'RE': 'FR',  # Reunion,                           French territory
+    'TF': 'FR',  # French Southern Territories,       French territory
+    'WF': 'FR',  # Wallis and Futuna,                 French territory
+    'YT': 'FR',  # Mayotte,                           French territory
+    'GI': 'GB',  # Gibraltar,                         British territory
+    'IM': 'GB',  # Isle of Man,                       British territory
+    'IO': 'GB',  # British Indian Ocean Territory,    British territory
+    'PN': 'GB',  # Pitcairn,                          British territory
+    'SH': 'GB',  # Saint Helena,                      British territory
+    'VG': 'GB',  # British Virgin Islands,            British territory
+    'BQ': 'NL',  # Bonaire,                           Dutch territory
+    'SX': 'NL',  # Sint Maarten,                      Dutch territory
+    'BV': 'NO',  # Bouvet Island,                     Norwegian territory
+    'SJ': 'NO',  # Svalbard and Jan Mayen,            Norwegian territory
+    'AS': 'US',  # American Samoa,                    American territory
+    'GU': 'US',  # Guam,                              American territory
+    'MP': 'US',  # Northern Mariana Islands,          American territory
+    'VI': 'US',  # US Virgin Islands,                 American territory
 }
 
+COUNTRY_ALIASES = {
+    # European Commission country code exceptions.
+    # Source: http://publications.europa.eu/code/pdf/370000en.htm#pays
+    'UK': 'GB',  # United Kingdom is known as 'GB' in ISO-3166
+    'EL': 'GR',  # 'EL' is the european version of Greece,
+}
 
-SUBDIVISION_ALIASES = {
+SUBDIVISION_COUNTRIES = {
     'CN-71': 'TW',  # Taiwan
     'CN-91': 'HK',  # Hong Kong
     'CN-92': 'MO',  # Macao
@@ -79,9 +134,6 @@ SUBDIVISION_ALIASES = {
     'FR-WF': 'WF',  # Wallis and Futuna
     'FR-YT': 'YT',  # Mayotte
     'NL-AW': 'AW',  # Aruba
-    'NL-BQ1': 'BQ-BO',  # Bonaire
-    'NL-BQ2': 'BQ-SA',  # Saba
-    'NL-BQ3': 'BQ-SE',  # Sint Eustatius
     'NL-CW': 'CW',  # Curaçao
     'NL-SX': 'SX',  # Sint Maarten
     'NO-21': 'SJ',  # Svalbard
@@ -94,12 +146,50 @@ SUBDIVISION_ALIASES = {
     'US-VI': 'VI',  # Virgin Islands, U.S.
 }
 
+SUBDIVISION_ALIASES = {
+    'NL-BQ1': 'BQ-BO',  # Bonaire
+    'NL-BQ2': 'BQ-SA',  # Saba
+    'NL-BQ3': 'BQ-SE',  # Sint Eustatius
+}
 
-# Build the reverse index of aliases defined above.
-REVERSE_MAPPING = {}
-for mapping in [COUNTRY_ALIASES, SUBDIVISION_ALIASES]:
-    for alias_code, target_code in mapping.items():
-        REVERSE_MAPPING.setdefault(target_code, set()).add(alias_code)
+RESERVED_COUNTRY_CODES = {
+    # Source:
+    # https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2#Exceptional_reservations
+    'DG': 'IO',  # Diego Garcia is part of the British Indian Ocean Territory
+    'FX': 'FR',  # France, Metropolitan
+    # European Commission country code exceptions.
+    # Source: http://publications.europa.eu/code/pdf/370000en.htm#pays
+    'EA': 'ES',  # 'EA' is the union of Ceuta and Melilla, Spanish territory
+}
+
+COUNTRY_ALIAS_TO_SUBDIVISION = {
+    'AC': 'SH-AC',  # Ascension Island
+    'CP': 'FR-CP',  # Clipperton Island
+    'IC': 'ES-CN',  # Canary Islands
+    'TA': 'SH-TA',  # Tristan da Cunha
+}
+
+
+def generate_mapping():
+    """Build the reverse index of aliases defined above.
+
+    :return: A dictionary containing aliases mapping.
+    """
+    mapping = {}
+    for reverse_mapping in [SUBDIVISION_COUNTRIES]:
+        for alias_code, target_code in reverse_mapping.items():
+            mapping.setdefault(target_code, set()).add(alias_code)
+
+    for straight_mapping in [RESERVED_COUNTRY_CODES,
+                             COUNTRY_ALIASES,
+                             SUBDIVISION_ALIASES,
+                             FOREIGN_TERRITORIES_MAPPING]:
+        for alias_code, target_code in straight_mapping.items():
+            mapping.setdefault(alias_code, set()).add(target_code)
+    return mapping
+
+
+REVERSE_MAPPING = generate_mapping()
 
 
 @cached(LRI())
@@ -120,7 +210,9 @@ def supported_country_codes():
     return set(chain(
         imap(attrgetter('alpha_2'), countries),
         # Include ISO and EC exceptions.
-        COUNTRY_ALIASES.keys()))
+        COUNTRY_ALIASES.keys(),
+        RESERVED_COUNTRY_CODES.keys(),
+        COUNTRY_ALIAS_TO_SUBDIVISION.keys()))
 
 
 @cached(LRI())
@@ -133,17 +225,45 @@ def supported_subdivision_codes():
     return set(imap(attrgetter('code'), subdivisions))
 
 
-def normalize_territory_code(territory_code, resolve_aliases=True):
-    """ Normalize any string into a territory code. """
+def normalize_territory_code(territory_code, resolve_aliases=True,
+                             resolve_top_country=False):
+    """Normalize any string into a territory code.
+
+    :param territory_code: The input string to normalize.
+    :param resolve_aliases: Trigger alias computation.
+    :param resolve_top_country: Trigger foreign country computation.
+    :return: The resolved territory code.
+    """
     territory_code = territory_code.strip().upper()
     if territory_code not in supported_territory_codes():
         raise ValueError(
             'Unrecognized {!r} territory code.'.format(territory_code))
+
+    # We resolve country aliases and subdivision aliases
+    # nevertheless since their keys does not exists in pycountry!
+    territory_code = RESERVED_COUNTRY_CODES.get(territory_code,
+                                                territory_code)
+    territory_code = COUNTRY_ALIASES.get(territory_code,
+                                         territory_code)
     if resolve_aliases:
-        territory_code = COUNTRY_ALIASES.get(territory_code, territory_code)
         territory_code = SUBDIVISION_ALIASES.get(
             territory_code, territory_code)
+        territory_code = SUBDIVISION_COUNTRIES.get(
+            territory_code, territory_code)
+    if resolve_top_country:
+        territory_code = territory_attachment(territory_code)
     return territory_code
+
+
+def territory_attachment(country_code):
+    """Returns the ISO-3166 alpha2 country_code of the country of which the
+    given country is part of.
+
+    :param country_code: The foreign territory to lookup.
+    :return: The main country of this foreign territory,
+    the input country_code if None.
+    """
+    return FOREIGN_TERRITORIES_MAPPING.get(country_code, country_code)
 
 
 def country_from_subdivision(subdivision_code):
@@ -156,7 +276,7 @@ def country_from_subdivision(subdivision_code):
     the later instead of the parent ISO 3166-2 top entry.
     """
     # Resolve subdivision alias.
-    code = SUBDIVISION_ALIASES.get(subdivision_code, subdivision_code)
+    code = SUBDIVISION_COUNTRIES.get(subdivision_code, subdivision_code)
 
     # We have a country code, return it right away.
     if code in supported_country_codes():
@@ -171,27 +291,30 @@ def country_from_subdivision(subdivision_code):
 
 
 def default_subdivision_code(country_code):
-    """ Return the default subdivision code of a country.
+    """Return the default subdivision code of a country.
 
     The result can be guessed only if there is a 1:1 mapping between a country
     code and a subdivision code.
+
+    :param country_code: Country code to find subdivision for.
+    :return: The subdivision key if found, None otherwise.
     """
     # Build the reverse index of the subdivision/country alias mapping.
     default_subdiv = {}
-    for subdiv_code, alias_code in SUBDIVISION_ALIASES.items():
+    for subdiv_code, alias_code in SUBDIVISION_COUNTRIES.items():
         # Skip non-country
         if len(alias_code) == 2:
             default_subdiv.setdefault(alias_code, set()).add(subdiv_code)
 
     # Include countries directly mapping to a subdivision.
-    for alias_code, subdiv_code in COUNTRY_ALIASES.items():
-        # Skip non-subdiv
-        if len(subdiv_code) > 3:
-            default_subdiv.setdefault(alias_code, set()).add(subdiv_code)
+    for alias_code, subdiv_code in COUNTRY_ALIAS_TO_SUBDIVISION.items():
+        default_subdiv.setdefault(alias_code, set()).add(subdiv_code)
 
     default_subdivisions = default_subdiv.get(country_code)
+
     if default_subdivisions and len(default_subdivisions) == 1:
         return default_subdivisions.pop()
+    return None
 
 
 def territory_children_codes(territory_code, include_self=False):
@@ -234,6 +357,9 @@ def territory_parents(territory_code, include_country=True):
     """
     tree = []
 
+    # Retrieving subdivision from alias to get full paternity
+    territory_code = COUNTRY_ALIAS_TO_SUBDIVISION.get(territory_code,
+                                                      territory_code)
     # If the provided territory code is a country, return it right away.
     territory_code = normalize_territory_code(territory_code)
     if territory_code in supported_country_codes():
@@ -269,7 +395,7 @@ def territory_parents_codes(territory_code, include_country=True):
         elif full_class_name == 'pycountry.db.Subdivision':
             yield territory.code
         else:
-            raise "Unrecognized {!r} territory.".format(territory)
+            raise ValueError("Unrecognized {!r} territory.".format(territory))
 
 
 def country_aliases(territory_code):
@@ -292,6 +418,9 @@ def country_aliases(territory_code):
         if not parent_code:
             parent_code = subdiv.country.alpha_2
         country_codes.update(country_aliases(parent_code))
+        # Adding subdivision's country alias
+        if territory_code in SUBDIVISION_COUNTRIES:
+            country_codes.update({SUBDIVISION_COUNTRIES.get(territory_code)})
 
     # Hunt for aliases
     for mapped_code in REVERSE_MAPPING.get(territory_code, []):