+++ CHANGES.rst
@@ -2,12 +2,14 @@ ChangeLog
 =========
 
 
-`1.3.6 (unreleased) <https://github.com/scaleway/postal-address/compare/v1.3.5...develop>`_
+`1.4.0 (unreleased) <https://github.com/scaleway/postal-address/compare/v1.3.5...develop>`_
 -------------------------------------------------------------------------------------------
 
 .. note:: This version is not yet released and is under active development.
 
-* No changes yet.
+* Unfreeze pycountry version and bump to ``pycountry >= 18.5.26``.
+* Refactor country aliases with better categories.
+* Some subdivisions names were updated, such as ``FR-O`` to ``FR-HDF``
 
 
 `1.3.5 (2017-10-03) <https://github.com/scaleway/postal-address/compare/v1.3.4...v1.3.5>`_